# openapi-redoc.yaml
openapi: 3.0.0
info:
  title: HarmonyUI V1 API
  description: Use these APIs to interact with the  HarmonyUI backend. You can find
    out more about the HarmonyUI project at htttps://github.com/MediaEase/HarmonyUI
  termsOfService: https:///github.com/MediaEase/HarmonyUI/tos
  contact:
    name: MediaEase
    url: htttps://github.com/MediaEase/MediaEase
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development server
security:
- Bearer: []
- apiKey: []
...
paths:
...
  /api/me/preferences:
    get:
      tags:
      - Profile
      summary: Get user preferences
      operationId: getPreferences
      responses:
        '200':
          description: Returns user preferences
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    patch:
      tags:
      - Profile
      summary: Update user preferences
      operationId: updatePreferences
      requestBody:
        content:
          application/hal+json:
            schema:
              $ref: '#/components/schemas/Preferences'
      responses:
        '200':
          description: User preferences updated
          content:
            application/hal+json:
              schema:
                type: object
                properties:
                  preferences:
                    type: string
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  /api/me/preferences/pin:
    patch:
      tags:
      - Profile
      summary: Update user pin
      operationId: updatePin
      requestBody:
        content:
          application/hal+json:
            schema:
              $ref: '#/components/schemas/PinnedApp'
      responses:
        '200':
          description: User pin updated
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PinnedApp'
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
...
  /api/me/preferences/widgets:
    patch:
      tags:
      - Profile
      summary: Update user widgets
      operationId: updateWidgets
      requestBody:
        content:
          application/hal+json:
            schema:
              $ref: '#/components/schemas/Widget'
      responses:
        '200':
          description: User widgets updated
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Widget'
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
...
  /api/widgets:
    get:
      tags:
      - Widgets
      summary: Get widgets
      operationId: getWidgets
      responses:
        '200':
          description: Returns widgets
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
    apiKey:
      type: apiKey
      description: 'Value: {apiKey}'
      in: header
      name: X-API-KEY
  responses:
    '204':
      description: No content - indicates the request was successful but there is
        no content to be returned in the response.
...
    '401':
      description: Unauthorized - Returned if there is no authorization header, or
        if the JWT token is expired.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                description: A message describing the error
                example: 'JWT validation failed: JWT is expired'
    '403':
      description: Forbidden - Returned if the user you are running as, doesn't have
        access to this end-point.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
          examples:
            '403':
              summary: An example of a 403 response object
              value:
                detailCode: 403 Forbidden
                trackingId: b21b1f7ce4da4d639f2c62a57171b427
                messages:
                - locale: en-US
                  localeOrigin: DEFAULT
                  text: The server understood the request but refuses to authorize
                    it.
...
    '422':
      description: Unprocessable Entity - Returned if the request was not processed
        due to validation errors.
    '500':
      description: Internal Server Error - Returned if there is an unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
          examples:
            '500':
              summary: An example of a 500 response object
              value:
                detailCode: 500.0 Internal Fault
                trackingId: b21b1f7ce4da4d639f2c62a57171b427
                messages:
                - locale: en-US
                  localeOrigin: DEFAULT
                  text: An internal fault occurred.
  schemas:
...
    LocaleOrigin:
      type: string
      enum:
      - DEFAULT
      - REQUEST
      description: An indicator of how the locale was selected. *DEFAULT* means the
        locale is the system default. *REQUEST* means the locale was selected from
        the request context (i.e., best match based on the *Accept-Language* header).
        Additional values may be added in the future without notice.
      example: DEFAULT
    ErrorMessageDto:
      type: object
      properties:
        locale:
          type: string
          description: The locale for the message text, a BCP 47 language tag.
          example: en-US
        localeOrigin:
          $ref: '#/components/schemas/LocaleOrigin'
        text:
          type: string
          description: Actual text of the error message in the indicated locale.
          example: The request was syntactically correct but its content is semantically
            invalid.
    ErrorResponseDto:
      type: object
      properties:
        detailCode:
          type: string
          description: Fine-grained error code providing more detail of the error.
          example: 400.1 Bad Request Content
        trackingId:
          type: string
          description: Unique tracking id for the error.
          example: e7eab60924f64aa284175b9fa3309599
        messages:
          type: array
          description: Generic localized reason for error
          items:
            $ref: '#/components/schemas/ErrorMessageDto'
        causes:
          type: array
          description: Plain-text descriptive reasons to provide additional detail
            to the text provided in the messages field
          items:
            $ref: '#/components/schemas/ErrorMessageDto'
...
    Preferences:
      type: object
      properties:
        pinnedApps:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
        display:
          type: string
        shell:
          type: string
        selectedWidgets:
          type: array
          items:
            type: string
        theme:
          type: string
    PinnedApp:
      type: object
      properties:
        pin:
          type: string
...
    Widget:
      type: object
      properties:
        widgets:
          type: array
          items:
            type: string
...